requirements.txt
django~=4.2.16
django-rest-framework
pip~=23.2.1
wheel~=0.41.2
asgiref~=3.8.1
sqlparse~=0.5.1
djangorestframework~=3.15.2
setuptools~=68.2.0
PyYAML~=6.0.2
attrs~=24.2.0
referencing~=0.35.1
jsonschema~=4.23.0
jsonschema-specifications~=2024.10.1
uritemplate~=4.1.1
inflection~=0.5.1
drf-spectacular~=0.27.2